[refactor] #153 가맹점 Controller에서 본사 Controller로 API 이동
