+++ module-app/app-api/src/main/java/com/chaing/api/controller/hq/HQOrderController.java
@@ -1,11 +1,13 @@
 package com.chaing.api.controller.hq;
 
 import com.chaing.api.security.principal.UserPrincipal;
+import com.chaing.domain.orders.dto.request.FranchiseOrderStatusUpdateRequest;
 import com.chaing.domain.orders.dto.request.HQOrderCreateRequest;
 import com.chaing.domain.orders.dto.request.HQOrderUpdateStatusRequest;
 import com.chaing.api.facade.factory.HQOrderFacade;
 import com.chaing.core.dto.ApiResponse;
 import com.chaing.domain.orders.dto.request.HQOrderUpdateRequest;
+import com.chaing.domain.orders.dto.response.FranchiseOrderStatusShippingPendingResponse;
 import com.chaing.domain.orders.dto.response.HQOrderCancelResponse;
 import com.chaing.domain.orders.dto.response.HQOrderCreateResponse;
 import com.chaing.domain.orders.dto.response.HQOrderDetailResponse;
@@ -108,5 +110,14 @@ public class HQOrderController {
         Long userId = userPrincipal.getId();
 
         return ResponseEntity.ok(ApiResponse.success(hqOrderFacade.create(userId, request)));
+    }
+
+    @Operation(summary = "발주 상태 수정", description = "차량 배정 후 해당 발주 상태를 배송 대기로 수정")
+    @PatchMapping("/shipping-pending")
+    @PreAuthorize("hasAnyRole('ADMIN', 'HQ')")
+    public ResponseEntity<ApiResponse<List<FranchiseOrderStatusShippingPendingResponse>>> updateOrderShippingPending(
+            @Valid @RequestBody List<FranchiseOrderStatusUpdateRequest> request
+    ) {
+        return ResponseEntity.ok(ApiResponse.success(hqOrderFacade.updateShippingPending(request)));
     }
 }
